# Use fused mul-add instructions where possible