--- a/fastapprox/src/fast/mod.rs
+++ b/fastapprox/src/fast/mod.rs
@@ -1,14 +1,14 @@
-use crate::faster;
 use crate::bits::*;
+use crate::faster;
 
 /// Base 2 logarithm.
 #[inline]
 pub fn log2(x: f32) -> f32 {
     let vx = to_bits(x);
     let mx = from_bits((vx & 0x007FFFFF_u32) | 0x3f000000);
-    let mut y = vx as f32;
-    y *= 1.1920928955078125e-7_f32;
-    y - 124.22551499_f32 - 1.498030302_f32 * mx - 1.72587999_f32 / (0.3520887068_f32 + mx)
+    let y = vx as f32;
+    y.mul_add(1.1920928955078125e-7_f32, -124.22551499_f32)
+        + mx.mul_add(-1.498030302_f32, -1.72587999_f32 / (0.3520887068_f32 + mx))
 }
 
 /// Natural logarithm.
@@ -24,7 +24,12 @@ pub fn pow2(p: f32) -> f32 {
     let clipp = if p < -126.0 { -126.0_f32 } else { p };
     let w = clipp as i32;
     let z = clipp - (w as f32) + offset;
-    let v = ((1 << 23) as f32 * (clipp + 121.2740575_f32 + 27.7280233_f32 / (4.84252568_f32 - z) - 1.49012907_f32 * z)) as u32;
+
+    let v = ((1 << 23) as f32
+        * z.mul_add(
+            -1.49012907_f32,
+            clipp + 121.2740575_f32 + 27.7280233_f32 / (4.84252568_f32 - z),
+        )) as u32;
     from_bits(v)
 }
 
@@ -54,7 +59,10 @@ pub fn ln_gamma(x: f32) -> f32 {
     let logterm = ln(x * (1.0_f32 + x) * (2.0_f32 + x));
     let xp3 = 3.0_f32 + x;
 
-    -2.081061466_f32 - x + 0.0833333_f32 / xp3 - logterm + (2.5_f32 + x) * ln(xp3)
+    (2.5_f32 + x).mul_add(
+        ln(xp3),
+        -2.081061466_f32 - x + 0.0833333_f32 / xp3 - logterm,
+    )
 }
 
 /// Digamma function.
@@ -65,7 +73,11 @@ pub fn digamma(x: f32) -> f32 {
     let twopx = 2.0_f32 + x;
     let logterm = ln(twopx);
 
-    (-48.0_f32 + x * (-157.0_f32 + x * (-127.0_f32 - 30.0_f32 * x))) / (12.0_f32 * x * (1.0_f32 + x) * twopx * twopx) + logterm
+    x.mul_add(
+        x.mul_add(x.mul_add(-30.0_f32, -127.0_f32), -157.0_f32),
+        -48.0_f32,
+    ) / (12.0_f32 * x * (1.0_f32 + x) * twopx * twopx)
+        + logterm
 }
 
 /// Complementary error function.
@@ -82,7 +94,10 @@ pub fn erfc(x: f32) -> f32 {
 
     v |= 0x80000000;
 
-    2.0_f32 / (1.0_f32 + pow2(K * x)) - A * x * (B * xquad - 1.0_f32) * faster::pow2(from_bits(v))
+    (-A * x * B.mul_add(xquad, -1.0_f32)).mul_add(
+        faster::pow2(from_bits(v)),
+        2.0_f32 / (1.0_f32 + pow2(K * x)),
+    )
 }
 
 /// Error function.
@@ -102,7 +117,10 @@ pub fn erf_inv(x: f32) -> f32 {
 
     let xsq = x * x;
 
-    INVK * log2((1.0_f32 + x) / (1.0_f32 - x)) + x * (A - B * xsq) / (C - D * xsq)
+    INVK.mul_add(
+        log2((1.0_f32 + x) / (1.0_f32 - x)),
+        x * (-B).mul_add(xsq, A) / (-D).mul_add(xsq, C),
+    )
 }
 
 /// Hyperbolic sine function.
@@ -128,11 +146,23 @@ pub fn tanh(p: f32) -> f32 {
 pub fn lambertw(x: f32) -> f32 {
     const THRESHOLD: f32 = 2.26445;
 
-    let c = if x < THRESHOLD { 1.546865557_f32 } else { 1.0_f32 };
-    let d = if x < THRESHOLD { 2.250366841_f32 } else { 0.0_f32 };
-    let a = if x < THRESHOLD { -0.737769969_f32 } else { 0.0_f32 };
+    let c = if x < THRESHOLD {
+        1.546865557_f32
+    } else {
+        1.0_f32
+    };
+    let d = if x < THRESHOLD {
+        2.250366841_f32
+    } else {
+        0.0_f32
+    };
+    let a = if x < THRESHOLD {
+        -0.737769969_f32
+    } else {
+        0.0_f32
+    };
 
-    let logterm = ln(c * x + d);
+    let logterm = ln(c.mul_add(x, d));
     let loglogterm = ln(logterm);
 
     let minusw = -a - logterm + loglogterm - loglogterm / logterm;
@@ -140,7 +170,10 @@ pub fn lambertw(x: f32) -> f32 {
     let xexpminusw = x * expminusw;
     let pexpminusw = xexpminusw - minusw;
 
-    (2.0_f32 * xexpminusw - minusw * (4.0_f32 * xexpminusw - minusw * pexpminusw)) / (2.0_f32 + pexpminusw * (2.0_f32 - minusw))
+    2.0_f32.mul_add(
+        xexpminusw,
+        -minusw * 4.0_f32.mul_add(xexpminusw, -minusw * pexpminusw),
+    ) / pexpminusw.mul_add(2.0_f32 - minusw, 2.0_f32)
 }
 
 /// Exponent of Lambert W function.
@@ -153,13 +186,14 @@ pub fn lambertwexpx(x: f32) -> f32 {
     let powarg = if x < K { A * (x - K) } else { 0.0_f32 };
 
     let logterm = ln(logarg);
-    let powterm = faster::pow2(powarg);  // don't need accuracy here
+    let powterm = faster::pow2(powarg); // don't need accuracy here
 
     let w = powterm * (logarg - logterm + logterm / logarg);
     let logw = ln(w);
     let p = x - logw;
 
-    w * (2.0_f32 + p + w * (3.0_f32 + 2.0_f32 * p)) / (2.0_f32 - p + w * (5.0_f32 + 2.0_f32 * w))
+    w * w.mul_add(p.mul_add(2.0_f32, 3.0_f32), 2.0_f32 + p)
+        / w.mul_add(w.mul_add(2.0_f32, 5.0_f32), 2.0_f32 - p)
 }
 
 /// Sine of a number in \[-π, π\], in radians.
@@ -177,14 +211,17 @@ pub fn sin(x: f32) -> f32 {
     let sign = v & 0x80000000;
     v &= 0x7FFFFFFF;
 
-    let qpprox = FOUROVERPI * x - FOUROVERPISQ * x * from_bits(v);
+    let qpprox = FOUROVERPI.mul_add(x, -FOUROVERPISQ * x * from_bits(v));
     let qpproxsq = qpprox * qpprox;
 
     p |= sign;
     r |= sign;
     s ^= sign;
 
-    Q * qpprox + qpproxsq * (from_bits(p) + qpproxsq * (from_bits(r) + qpproxsq * from_bits(s)))
+    Q.mul_add(
+        qpprox,
+        qpproxsq * qpproxsq.mul_add(qpproxsq.mul_add(from_bits(s), from_bits(r)), from_bits(p)),
+    )
 }
 
 /// Sine in radians.
@@ -197,7 +234,7 @@ pub fn sinfull(x: f32) -> f32 {
 
     let k: u32 = (x * INVTWOPI) as u32;
     let half = if x < 0_f32 { -0.5_f32 } else { 0.5_f32 };
-    sin((half + (k as f32)) * TWOPI - x)
+    sin((half + (k as f32)).mul_add(TWOPI, -x))
 }
 
 /// Cosine of a number in \[-π, π\], in radians.
@@ -206,7 +243,7 @@ pub fn sinfull(x: f32) -> f32 {
 ///
 /// ```
 /// assert_eq!(f32::cos(1.0), 0.5403023);
-/// assert_eq!(fastapprox::fast::cos(1.0), 0.5402951);
+/// assert_eq!(fastapprox::fast::cos(1.0), 0.54029506);
 /// ```
 #[inline]
 pub fn cos(x: f32) -> f32 {
@@ -224,7 +261,7 @@ pub fn cos(x: f32) -> f32 {
 ///
 /// ```
 /// assert_eq!(f32::cos(10.0), -0.8390715);
-/// assert_eq!(fastapprox::fast::cosfull(10.0), -0.83907986);
+/// assert_eq!(fastapprox::fast::cosfull(10.0), -0.83908);
 /// ```
 #[inline]
 pub fn cosfull(x: f32) -> f32 {
@@ -249,7 +286,7 @@ pub fn tanfull(x: f32) -> f32 {
 
     let k: u32 = (x * INVTWOPI) as u32;
     let half = if x < 0_f32 { -0.5_f32 } else { 0.5_f32 };
-    let xnew = x - (half + k as f32) * TWOPI;
+    let xnew = (half + k as f32).mul_add(-TWOPI, x);
 
     sin(xnew) / cos(xnew)
 }