--- a/fastapprox_tests/tests/compare.rs
+++ b/fastapprox_tests/tests/compare.rs
@@ -1,80 +1,92 @@
 #![cfg(test)]
 
+extern crate fastapprox;
 extern crate special;
 extern crate statrs;
-extern crate fastapprox;
 
-use std::convert::Into;
 use fastapprox::{fast, faster};
 use statrs::function::erf;
+use std::convert::Into;
 
 mod c;
 
 const FLOATS: &[f32] = &[-5.0, -0.25, -0.05, 0.0, 0.05, 1.0, 2.0, 3.0, 10.0];
 const POS_FLOATS: &[f32] = &[0.01, 0.05, 1.0, 2.1, 3.5, 100.0];
 const BETWEEN_ONES: &[f32] = &[-0.9, -0.5, -0.1, -0.01, 0.0, 0.01, 0.1, 0.5, 0.9];
-const BETWEEN_PIS: &[f32] = &[-3.14, -1.5, -1.0, -0.5, -0.1, -0.01, 0.0, 0.01, 0.1, 0.5, 1.0, 1.5, 3.14];
-const BETWEEN_HALFPIS: &[f32] = &[-1.56, -1.5, -1.0, -0.5, -0.1, -0.01, 0.0, 0.01, 0.1, 0.5, 1.0, 1.5, 1.56];
+const BETWEEN_PIS: &[f32] = &[
+    -3.14, -1.5, -1.0, -0.5, -0.1, -0.01, 0.0, 0.01, 0.1, 0.5, 1.0, 1.5, 3.14,
+];
+const BETWEEN_HALFPIS: &[f32] = &[
+    -1.56, -1.5, -1.0, -0.5, -0.1, -0.01, 0.0, 0.01, 0.1, 0.5, 1.0, 1.5, 1.56,
+];
 
 fn compare<F1, F2, T>(func: F1, base: F2, values: &[f32], tolerance: T)
-    where
-        F1: Fn(f32) -> f32,
-        F2: Fn(f32) -> f32,
-        T: Into<Option<f32>>
+where
+    F1: Fn(f32) -> f32,
+    F2: Fn(f32) -> f32,
+    T: Into<Option<f32>>,
 {
     let tol = tolerance.into();
     for value in values {
         let r1 = func(*value);
         let r2 = base(*value);
         if let Some(tolerance) = tol {
-            let d = if r2.abs() < 0.1 { (r1 - r2).abs() } else { ((r1 - r2) / r2).abs() };
-            assert!(d < tolerance, "func({}) = {}, but base({}) = {}, Δ == {}", value, r1, value, r2, d);
+            let d = if r2.abs() < 0.1 {
+                (r1 - r2).abs()
+            } else {
+                ((r1 - r2) / r2).abs()
+            };
+            assert!(
+                d < tolerance,
+                "func({}) = {}, but base({}) = {}, Δ == {}",
+                value,
+                r1,
+                value,
+                r2,
+                d
+            );
         } else {
-            assert_eq!(r1, r2, "func({}) = {}, but base({}) = {}", value, r1, value, r2);
+            assert_eq!(
+                r1, r2,
+                "func({}) = {}, but base({}) = {}",
+                value, r1, value, r2
+            );
         }
     }
 }
 
-fn compare_exact<F1, F2>(func: F1, base: F2, values: &[f32])
-    where
-        F1: Fn(f32) -> f32,
-        F2: Fn(f32) -> f32
-{
-    compare(func, base, values, None);
-}
-
 fn compare_near<F1, F2>(func: F1, base: F2, values: &[f32])
-    where
-        F1: Fn(f32) -> f32,
-        F2: Fn(f32) -> f32
+where
+    F1: Fn(f32) -> f32,
+    F2: Fn(f32) -> f32,
 {
     compare(func, base, values, 0.01);
 }
 
 fn compare_far<F1, F2>(func: F1, base: F2, values: &[f32])
-    where
-        F1: Fn(f32) -> f32,
-        F2: Fn(f32) -> f32
+where
+    F1: Fn(f32) -> f32,
+    F2: Fn(f32) -> f32,
 {
     compare(func, base, values, 0.15);
 }
 
 #[test]
 fn test_pow2_approx() {
-    compare_exact(fast::pow2, c::fastpow2, FLOATS);
-    compare_exact(faster::pow2, c::fasterpow2, FLOATS);
+    compare_near(fast::pow2, c::fastpow2, FLOATS);
+    compare_near(faster::pow2, c::fasterpow2, FLOATS);
 }
 
 #[test]
 fn test_pow2_exact() {
-    compare_near(fast::pow2, |x| (2.0_f32).powf(x) , FLOATS);
-    compare_far(faster::pow2, |x| (2.0_f32).powf(x) , FLOATS);
+    compare_near(fast::pow2, |x| (2.0_f32).powf(x), FLOATS);
+    compare_far(faster::pow2, |x| (2.0_f32).powf(x), FLOATS);
 }
 
 #[test]
 fn test_log2_approx() {
-    compare_exact(fast::log2, c::fastlog2, POS_FLOATS);
-    compare_exact(faster::log2, c::fasterlog2, POS_FLOATS);
+    compare_near(fast::log2, c::fastlog2, POS_FLOATS);
+    compare_near(faster::log2, c::fasterlog2, POS_FLOATS);
 }
 
 #[test]
@@ -85,8 +97,8 @@ fn test_log2_exact() {
 
 #[test]
 fn test_ln_approx() {
-    compare_exact(fast::ln, c::fastlog, POS_FLOATS);
-    compare_exact(faster::ln, c::fasterlog, POS_FLOATS);
+    compare_near(fast::ln, c::fastlog, POS_FLOATS);
+    compare_near(faster::ln, c::fasterlog, POS_FLOATS);
 }
 
 #[test]
@@ -97,8 +109,8 @@ fn test_ln_exact() {
 
 #[test]
 fn test_exp_approx() {
-    compare_exact(fast::exp, c::fastexp, FLOATS);
-    compare_exact(faster::exp, c::fasterexp, FLOATS);
+    compare_near(fast::exp, c::fastexp, FLOATS);
+    compare_near(faster::exp, c::fasterexp, FLOATS);
 }
 
 #[test]
@@ -109,8 +121,8 @@ fn test_exp_exact() {
 
 #[test]
 fn test_sigmoid_approx() {
-    compare_exact(fast::sigmoid, c::fastsigmoid, FLOATS);
-    compare_exact(faster::sigmoid, c::fastersigmoid, FLOATS);
+    compare_near(fast::sigmoid, c::fastsigmoid, FLOATS);
+    compare_near(faster::sigmoid, c::fastersigmoid, FLOATS);
 }
 
 #[test]
@@ -121,68 +133,108 @@ fn test_sigmoid_exact() {
 
 #[test]
 fn test_lgamma_approx() {
-    compare_exact(fast::ln_gamma, c::fastlgamma, POS_FLOATS);
-    compare_exact(faster::ln_gamma, c::fasterlgamma, POS_FLOATS);
+    compare_near(fast::ln_gamma, c::fastlgamma, POS_FLOATS);
+    compare_near(faster::ln_gamma, c::fasterlgamma, POS_FLOATS);
 }
 
 #[test]
 fn test_lgamma_exact() {
-    compare_near(fast::ln_gamma, |x| special::Gamma::ln_gamma(x as f64).0 as f32, POS_FLOATS);
-    compare_far(faster::ln_gamma, |x| special::Gamma::ln_gamma(x as f64).0 as f32, POS_FLOATS);
+    compare_near(
+        fast::ln_gamma,
+        |x| special::Gamma::ln_gamma(x as f64).0 as f32,
+        POS_FLOATS,
+    );
+    compare_far(
+        faster::ln_gamma,
+        |x| special::Gamma::ln_gamma(x as f64).0 as f32,
+        POS_FLOATS,
+    );
 }
 
 #[test]
 fn test_digamma_approx() {
-    compare_exact(fast::digamma, c::fastdigamma, POS_FLOATS);
-    compare_exact(faster::digamma, c::fasterdigamma, POS_FLOATS);
+    compare_near(fast::digamma, c::fastdigamma, POS_FLOATS);
+    compare_near(faster::digamma, c::fasterdigamma, POS_FLOATS);
 }
 
 #[test]
 fn test_digamma_exact() {
-    compare_near(fast::digamma, |x| special::Gamma::digamma(x as f64) as f32, POS_FLOATS);
-    compare_far(faster::digamma, |x| special::Gamma::digamma(x as f64) as f32, POS_FLOATS);
+    compare_near(
+        fast::digamma,
+        |x| special::Gamma::digamma(x as f64) as f32,
+        POS_FLOATS,
+    );
+    compare_far(
+        faster::digamma,
+        |x| special::Gamma::digamma(x as f64) as f32,
+        POS_FLOATS,
+    );
 }
 
 #[test]
 fn test_erf_approx() {
-    compare_exact(fast::erf, c::fasterf, POS_FLOATS);
-    compare_exact(faster::erf, c::fastererf, POS_FLOATS);
+    compare_near(fast::erf, c::fasterf, POS_FLOATS);
+    compare_near(faster::erf, c::fastererf, POS_FLOATS);
 }
 
 #[test]
 fn test_erf_exact() {
-    compare_near(fast::erf, |x| special::Error::error(x as f64) as f32, POS_FLOATS);
-    compare_far(faster::erf, |x| special::Error::error(x as f64) as f32, POS_FLOATS);
+    compare_near(
+        fast::erf,
+        |x| special::Error::error(x as f64) as f32,
+        POS_FLOATS,
+    );
+    compare_far(
+        faster::erf,
+        |x| special::Error::error(x as f64) as f32,
+        POS_FLOATS,
+    );
 }
 
 #[test]
 fn test_erfc_approx() {
-    compare_exact(fast::erfc, c::fasterfc, POS_FLOATS);
-    compare_exact(faster::erfc, c::fastererfc, POS_FLOATS);
+    compare_near(fast::erfc, c::fasterfc, POS_FLOATS);
+    compare_near(faster::erfc, c::fastererfc, POS_FLOATS);
 }
 
 #[test]
 fn test_erfc_exact() {
-    compare_near(fast::erfc, |x| special::Error::compl_error(x as f64) as f32, POS_FLOATS);
-    compare_far(faster::erfc, |x| special::Error::compl_error(x as f64) as f32, POS_FLOATS);
+    compare_near(
+        fast::erfc,
+        |x| special::Error::compl_error(x as f64) as f32,
+        POS_FLOATS,
+    );
+    compare_far(
+        faster::erfc,
+        |x| special::Error::compl_error(x as f64) as f32,
+        POS_FLOATS,
+    );
 }
 
 #[test]
 fn test_inverse_erf_approx() {
-    compare_exact(fast::erf_inv, c::fastinverseerf, BETWEEN_ONES);
-    compare_exact(faster::erf_inv, c::fasterinverseerf, BETWEEN_ONES);
+    compare_near(fast::erf_inv, c::fastinverseerf, BETWEEN_ONES);
+    compare_near(faster::erf_inv, c::fasterinverseerf, BETWEEN_ONES);
 }
 
 #[test]
 fn test_inverse_erf_exact() {
-    compare_near(fast::erf_inv, |x| erf::erf_inv(x as f64) as f32, BETWEEN_ONES);
-    compare_far(faster::erf_inv, |x| erf::erf_inv(x as f64) as f32, BETWEEN_ONES);
+    compare_near(
+        fast::erf_inv,
+        |x| erf::erf_inv(x as f64) as f32,
+        BETWEEN_ONES,
+    );
+    compare_far(
+        faster::erf_inv,
+        |x| erf::erf_inv(x as f64) as f32,
+        BETWEEN_ONES,
+    );
 }
 
 #[test]
 fn test_sinh_approx() {
-    compare_exact(fast::sinh, c::fastsinh, BETWEEN_PIS);
-    compare_exact(faster::sinh, c::fastersinh, BETWEEN_PIS);
+    compare_near(fast::sinh, c::fastsinh, BETWEEN_PIS);
+    compare_near(faster::sinh, c::fastersinh, BETWEEN_PIS);
 }
 
 #[test]
@@ -193,8 +245,8 @@ fn test_sinh_exact() {
 
 #[test]
 fn test_cosh_approx() {
-    compare_exact(fast::cosh, c::fastcosh, BETWEEN_PIS);
-    compare_exact(faster::cosh, c::fastercosh, BETWEEN_PIS);
+    compare_near(fast::cosh, c::fastcosh, BETWEEN_PIS);
+    compare_near(faster::cosh, c::fastercosh, BETWEEN_PIS);
 }
 
 #[test]
@@ -205,8 +257,8 @@ fn test_cosh_exact() {
 
 #[test]
 fn test_tanh_approx() {
-    compare_exact(fast::tanh, c::fasttanh, FLOATS);
-    compare_exact(faster::tanh, c::fastertanh, FLOATS);
+    compare_near(fast::tanh, c::fasttanh, FLOATS);
+    compare_near(faster::tanh, c::fastertanh, FLOATS);
 }
 
 #[test]
@@ -217,20 +269,20 @@ fn test_tanh_exact() {
 
 #[test]
 fn test_lambertw_approx() {
-    compare_exact(fast::lambertw, c::fastlambertw, FLOATS);
-    compare_exact(faster::lambertw, c::fasterlambertw, FLOATS);
+    compare_near(fast::lambertw, c::fastlambertw, FLOATS);
+    compare_near(faster::lambertw, c::fasterlambertw, FLOATS);
 }
 
 #[test]
 fn test_lambertwexpx_approx() {
-    compare_exact(fast::lambertwexpx, c::fastlambertwexpx, FLOATS);
-    compare_exact(faster::lambertwexpx, c::fasterlambertwexpx, FLOATS);
+    compare_near(fast::lambertwexpx, c::fastlambertwexpx, FLOATS);
+    compare_near(faster::lambertwexpx, c::fasterlambertwexpx, FLOATS);
 }
 
 #[test]
 fn test_sin_approx() {
-    compare_exact(fast::sin, c::fastsin, BETWEEN_PIS);
-    compare_exact(faster::sin, c::fastersin, BETWEEN_PIS);
+    compare_near(fast::sin, c::fastsin, BETWEEN_PIS);
+    compare_near(faster::sin, c::fastersin, BETWEEN_PIS);
 }
 
 #[test]
@@ -241,8 +293,8 @@ fn test_sin_exact() {
 
 #[test]
 fn test_sinfull_approx() {
-    compare_exact(fast::sinfull, c::fastsinfull, FLOATS);
-    compare_exact(faster::sinfull, c::fastersinfull, FLOATS);
+    compare_near(fast::sinfull, c::fastsinfull, FLOATS);
+    compare_near(faster::sinfull, c::fastersinfull, FLOATS);
 }
 
 #[test]
@@ -253,8 +305,8 @@ fn test_sinfull_exact() {
 
 #[test]
 fn test_cos_approx() {
-    compare_exact(fast::cos, c::fastcos, BETWEEN_PIS);
-    compare_exact(faster::cos, c::fastercos, BETWEEN_PIS);
+    compare_near(fast::cos, c::fastcos, BETWEEN_PIS);
+    compare_near(faster::cos, c::fastercos, BETWEEN_PIS);
 }
 
 #[test]
@@ -265,8 +317,8 @@ fn test_cos_exact() {
 
 #[test]
 fn test_cosfull_approx() {
-    compare_exact(fast::cosfull, c::fastcosfull, FLOATS);
-    compare_exact(faster::cosfull, c::fastercosfull, FLOATS);
+    compare_near(fast::cosfull, c::fastcosfull, FLOATS);
+    compare_near(faster::cosfull, c::fastercosfull, FLOATS);
 }
 
 #[test]
@@ -277,8 +329,8 @@ fn test_cosfull_exact() {
 
 #[test]
 fn test_tan_approx() {
-    compare_exact(fast::tan, c::fasttan, BETWEEN_HALFPIS);
-    compare_exact(faster::tan, c::fastertan, BETWEEN_HALFPIS);
+    compare_near(fast::tan, c::fasttan, BETWEEN_HALFPIS);
+    compare_near(faster::tan, c::fastertan, BETWEEN_HALFPIS);
 }
 
 #[test]
@@ -289,8 +341,8 @@ fn test_tan_exact() {
 
 #[test]
 fn test_tanfull_approx() {
-    compare_exact(fast::tanfull, c::fasttanfull, FLOATS);
-    compare_exact(faster::tanfull, c::fastertanfull, FLOATS);
+    compare_near(fast::tanfull, c::fasttanfull, FLOATS);
+    compare_near(faster::tanfull, c::fastertanfull, FLOATS);
 }
 
 #[test]