@loony-bean 2023-08-27
Hi @shssoichiro! First of all, thank you for this contribution. It's amazing that you were able to make these computations even faster w/o changing the principles.

But I'd suggest putting the new code using `f32::mul_add` into separate module to preserve full parity with the original C library for `fast` and `faster` modules. We can call the new module `fused` (because why not). If you do not have time for implementing it yourself, I can take it from here.

Thanks again for bringing this PR, and sorry that it took me so long to review it.